refactor(xinfa): modularize compiler internals
