+++ crates/xinfa/extraction-manifest.json
@@ -5,6 +5,7 @@
     "Cargo.lock",
     "Cargo.toml",
     "LICENSE",
+    "ARCHITECTURE.md",
     "README.md",
     "boundary.contract.json",
     "rust-toolchain.toml",
@@ -84,12 +85,18 @@
     "schema/task-chart-v1.schema.json",
     "src/atlas.rs",
     "src/cli.rs",
+    "src/command.rs",
     "src/engine.rs",
     "src/episode.rs",
+    "src/episode/admission.rs",
     "src/lib.rs",
     "src/main.rs",
     "src/native_io.rs",
     "src/pack.rs",
+    "src/pack/impact.rs",
+    "src/pack/source.rs",
+    "src/pack/verification.rs",
+    "src/project_validation.rs",
     "src/projection.rs",
     "src/resolver.rs",
     "src/semantic_project.rs",
