+++ crates/xinfa/src/pack.rs
@@ -6,10 +6,21 @@ use std::collections::{BTreeMap, BTreeSet, VecDeque};
 #[cfg(test)]
 use std::fs;
 #[cfg(test)]
+use std::path::Path;
+#[cfg(test)]
 use std::path::PathBuf;
-use std::path::{Component, Path};
 
 use super::{compile, digest, normalized, parse, stable_json, validate, visibility_rank};
+
+mod impact;
+mod source;
+mod verification;
+
+pub use impact::impact_between_values;
+use source::collect_inventory;
+#[cfg(test)]
+use verification::verify_value;
+pub use verification::{inspect_pack_value, verify_pack_artifacts};
 
 const MAX_SOURCE_BYTES: u64 = 4 * 1024 * 1024;
 
@@ -128,178 +139,6 @@ fn compile_receipt(
         "diagnostics": diagnostics_value(diagnostics),
         "compiler": {"product": "xinfa", "version": env!("CARGO_PKG_VERSION")},
     })
-}
-
-fn sensitive_path(relative: &str) -> bool {
-    relative.split('/').any(|part| {
-        part == ".git"
-            || part == ".private"
-            || part == ".env"
-            || part.starts_with(".env.")
-            || part.eq_ignore_ascii_case("secrets")
-            || part.eq_ignore_ascii_case("credentials.json")
-    })
-}
-
-fn generated_projection_path(relative: &str) -> bool {
-    relative == ".xinfa/generated" || relative.starts_with(".xinfa/generated/")
-}
-
-fn checked_source(
-    repository: &dyn RepositorySource,
-    relative: &str,
-) -> Result<(Vec<u8>, u64), PackDiagnostic> {
-    if generated_projection_path(relative) {
-        return Err(PackDiagnostic::error(
-            "generated-projection-input",
-            relative,
-            "generated Xinfa projections are derived data and cannot enter a provider; explicitly accept content into a managed source path and compile a successor cut",
-        ));
-    }
-    if sensitive_path(relative) {
-        return Err(PackDiagnostic::error(
-            "sensitive-path",
-            relative,
-            "sensitive path classes cannot enter a context pack",
-        ));
-    }
-    for component in Path::new(relative).components() {
-        let Component::Normal(_) = component else {
-            return Err(PackDiagnostic::error(
-                "invalid-path",
-                relative,
-                "source path must remain repository relative",
-            ));
-        };
-    }
-    let bytes = repository
-        .read(relative)
-        .map_err(|error| PackDiagnostic::error(error.code, relative, error.message))?;
-    if bytes.len() as u64 > MAX_SOURCE_BYTES {
-        return Err(PackDiagnostic::error(
-            "source-too-large",
-            relative,
-            format!("declared source exceeds the {MAX_SOURCE_BYTES} byte v1 limit"),
-        ));
-    }
-    let size = bytes.len() as u64;
-    Ok((bytes, size))
-}
-
-fn provider_inventory_root(entries: &[Value]) -> String {
-    digest(&Value::Array(
-        entries
-            .iter()
-            .map(|entry| {
-                json!({
-                    "path": entry["path"],
-                    "contentRoot": entry["contentRoot"],
-                    "size": entry["size"],
-                })
-            })
-            .collect(),
-    ))
-}
-
-fn collect_inventory(
-    project: &Value,
-    repository: &dyn RepositorySource,
-    visibility: &str,
-) -> (Vec<Value>, Vec<PackDiagnostic>) {
-    let mut inventory = Vec::new();
-    let mut diagnostics = Vec::new();
-    for provider in project["providers"]
-        .as_array()
-        .expect("validated providers")
-    {
-        let provider_visibility = provider["visibility"].as_str().expect("visibility");
-        if visibility_rank(provider_visibility) > visibility_rank(visibility) {
-            continue;
-        }
-        let id = provider["id"].as_str().expect("provider id");
-        let kind = provider["kind"].as_str().expect("provider kind");
-        if kind != "exact-file-manifest" {
-            diagnostics.push(
-                PackDiagnostic::error(
-                    "unsupported-provider",
-                    format!("/providers/{id}"),
-                    "repository pack v1 accepts only exact-file-manifest providers",
-                )
-                .with_provenance(json!({"provider": id, "kind": kind})),
-            );
-            continue;
-        }
-        let mut provider_entries = Vec::new();
-        for path in provider["paths"].as_array().expect("provider paths") {
-            let relative = path.as_str().expect("validated path");
-            match checked_source(repository, relative) {
-                Ok((bytes, size)) => match String::from_utf8(bytes) {
-                    Ok(content) => provider_entries.push(json!({
-                        "path": relative,
-                        "contentRoot": byte_digest(content.as_bytes()),
-                        "size": size,
-                        "encoding": "utf-8",
-                        "content": content,
-                    })),
-                    Err(_) => diagnostics.push(
-                        PackDiagnostic::error(
-                            "unsupported-encoding",
-                            relative,
-                            "repository pack v1 accepts only UTF-8 source units",
-                        )
-                        .with_provenance(json!({"provider": id})),
-                    ),
-                },
-                Err(error) => diagnostics.push(error.with_provenance(json!({"provider": id}))),
-            }
-        }
-        provider_entries.sort_by(|left, right| left["path"].as_str().cmp(&right["path"].as_str()));
-        let actual = provider_inventory_root(&provider_entries);
-        let expected = provider["revision"].as_str().expect("provider revision");
-        if actual != expected {
-            diagnostics.push(
-                PackDiagnostic::error(
-                    "provider-drift",
-                    format!("/providers/{id}/revision"),
-                    format!(
-                        "declared provider revision {expected} does not match observed {actual}"
-                    ),
-                )
-                .with_provenance(json!({"provider": id, "expected": expected, "observed": actual})),
-            );
-        }
-        for entry in provider_entries {
-            inventory.push(json!({
-                "provider": id,
-                "visibility": provider_visibility,
-                "path": entry["path"],
-                "contentRoot": entry["contentRoot"],
-                "size": entry["size"],
-                "encoding": entry["encoding"],
-                "content": entry["content"],
-            }));
-        }
-    }
-    inventory.sort_by(|left, right| {
-        (left["provider"].as_str(), left["path"].as_str())
-            .cmp(&(right["provider"].as_str(), right["path"].as_str()))
-    });
-    let mut owners = BTreeMap::new();
-    for item in &inventory {
-        let path = item["path"].as_str().expect("inventory path");
-        let provider = item["provider"].as_str().expect("inventory provider");
-        if let Some(previous) = owners.insert(path, provider) {
-            diagnostics.push(
-                PackDiagnostic::error(
-                    "duplicate-source-owner",
-                    path,
-                    "one repository path must not be acquired by multiple providers",
-                )
-                .with_provenance(json!({"providers": [previous, provider]})),
-            );
-        }
-    }
-    (inventory, diagnostics)
 }
 
 fn complete_routes(
@@ -673,468 +512,6 @@ pub fn compile_repository_pack_from_source(
         artifacts: Some(artifacts),
         receipt,
     })
-}
-
-fn pack_core(pack: &Value) -> Result<Value, String> {
-    let mut core = pack.clone();
-    core.pointer_mut("/roots")
-        .and_then(Value::as_object_mut)
-        .and_then(|roots| roots.remove("pack"))
-        .ok_or_else(|| "pack is missing roots.pack".to_owned())?;
-    Ok(core)
-}
-
-fn verify_value(pack: &Value) -> Vec<PackDiagnostic> {
-    let mut diagnostics = Vec::new();
-    if pack["schema"] != "xinfa.context-pack/v1" {
-        diagnostics.push(PackDiagnostic::error(
-            "pack-schema",
-            "/schema",
-            "must be xinfa.context-pack/v1",
-        ));
-        return diagnostics;
-    }
-    let expected = pack
-        .pointer("/roots/pack")
-        .and_then(Value::as_str)
-        .unwrap_or_default();
-    match pack_core(pack) {
-        Ok(core) if digest(&core) != expected => diagnostics.push(PackDiagnostic::error(
-            "pack-root",
-            "/roots/pack",
-            "pack root does not match canonical pack content",
-        )),
-        Err(message) => {
-            diagnostics.push(PackDiagnostic::error("pack-root", "/roots/pack", message))
-        }
-        _ => {}
-    }
-    if let Some(inventory) = pack["inventory"].as_array() {
-        for item in inventory {
-            let path = item["path"].as_str().unwrap_or_default();
-            let content = item["content"].as_str();
-            if item["encoding"] != "utf-8" || content.is_none() {
-                diagnostics.push(PackDiagnostic::error(
-                    "source-encoding",
-                    format!("/inventory/{path}"),
-                    "source unit must contain UTF-8 payload",
-                ));
-                continue;
-            }
-            let bytes = content.expect("checked content").as_bytes();
-            if byte_digest(bytes) != item["contentRoot"].as_str().unwrap_or_default()
-                || item["size"].as_u64() != Some(bytes.len() as u64)
-            {
-                diagnostics.push(PackDiagnostic::error(
-                    "source-content-root",
-                    format!("/inventory/{path}"),
-                    "source payload does not match its content root and size",
-                ));
-            }
-        }
-    } else {
-        diagnostics.push(PackDiagnostic::error(
-            "source-inventory",
-            "/inventory",
-            "pack inventory must be an array",
-        ));
-    }
-    for (field, value) in [
-        ("source", &pack["inventory"]),
-        ("coverage", &pack["coverage"]),
-        ("cut", &pack["cut"]),
-    ] {
-        let expected = pack["roots"][field].as_str().unwrap_or_default();
-        if digest(value) != expected {
-            diagnostics.push(PackDiagnostic::error(
-                "component-root",
-                format!("/roots/{field}"),
-                format!("{field} root does not match content"),
-            ));
-        }
-    }
-    let authority = json!({"cut": pack["cut"], "nodes": pack["nodes"], "edges": pack["edges"]});
-    if digest(&authority) != pack["roots"]["authority"].as_str().unwrap_or_default() {
-        diagnostics.push(PackDiagnostic::error(
-            "component-root",
-            "/roots/authority",
-            "authority root does not match content",
-        ));
-    }
-    let policy = json!({"policies": pack["policies"], "visibility": pack["visibility"], "routes": pack["routes"]});
-    if digest(&policy) != pack["roots"]["policy"].as_str().unwrap_or_default() {
-        diagnostics.push(PackDiagnostic::error(
-            "component-root",
-            "/roots/policy",
-            "policy root does not match content",
-        ));
-    }
-    let mut groups: BTreeMap<&str, Vec<&Value>> = BTreeMap::new();
-    let nodes: BTreeMap<&str, &Value> = pack["nodes"]
-        .as_array()
-        .map(Vec::as_slice)
-        .unwrap_or(&[])
-        .iter()
-        .filter_map(|node| Some((node["id"].as_str()?, node)))
-        .collect();
-    if let Some(routes) = pack["routes"].as_array() {
-        for route in routes {
-            let selected: Vec<Value> = route["nodes"]
-                .as_array()
-                .map(Vec::as_slice)
-                .unwrap_or(&[])
-                .iter()
-                .filter_map(|id| {
-                    let id = id.as_str()?;
-                    let node = nodes.get(id)?;
-                    Some(json!({
-                        "id": id,
-                        "revision": node["revision"],
-                        "status": node["verification"]["status"],
-                    }))
-                })
-                .collect();
-            if selected.len() != route["nodes"].as_array().map(Vec::len).unwrap_or_default()
-                || digest(&Value::Array(selected))
-                    != route["authorityRoot"].as_str().unwrap_or_default()
-            {
-                diagnostics.push(PackDiagnostic::error(
-                    "route-authority-root",
-                    format!("/routes/{}", route["id"].as_str().unwrap_or_default()),
-                    "route authority root does not match selected nodes",
-                ));
-            }
-            let mut source_route = route.clone();
-            if let Some(object) = source_route.as_object_mut() {
-                object.remove("authorityRoot");
-                object.remove("routeRoot");
-                object.remove("status");
-            }
-            if digest(&source_route) != route["routeRoot"].as_str().unwrap_or_default() {
-                diagnostics.push(PackDiagnostic::error(
-                    "route-root",
-                    format!("/routes/{}", route["id"].as_str().unwrap_or_default()),
-                    "route root does not match route declaration",
-                ));
-            }
-            let expected_status = if route["nodes"]
-                .as_array()
-                .map(Vec::as_slice)
-                .unwrap_or(&[])
-                .iter()
-                .filter_map(Value::as_str)
-                .filter_map(|id| nodes.get(id))
-                .any(|node| {
-                    matches!(
-                        node["verification"]["status"].as_str(),
-                        Some("stale" | "invalidated")
-                    )
-                }) {
-                "stale"
-            } else {
-                "current"
-            };
-            if route["status"] != expected_status {
-                diagnostics.push(PackDiagnostic::error(
-                    "route-status",
-                    format!("/routes/{}", route["id"].as_str().unwrap_or_default()),
-                    "route status does not match selected node status",
-                ));
-            }
-            groups
-                .entry(route["parityGroup"].as_str().unwrap_or_default())
-                .or_default()
-                .push(route);
-        }
-    }
-    for (group, routes) in groups {
-        if routes.len() != 2
-            || routes
-                .iter()
-                .map(|route| route["audience"].as_str())
-                .collect::<BTreeSet<_>>()
-                .len()
-                != 2
-            || routes[0]["authorityRoot"] != routes[1]["authorityRoot"]
-            || routes[0]["status"] != routes[1]["status"]
-        {
-            diagnostics.push(PackDiagnostic::error(
-                "route-parity",
-                "/routes",
-                format!("parity group {group} does not preserve dual-first authority"),
-            ));
-        }
-    }
-    if let (Some(nodes), Some(edges), Some(routes)) = (
-        pack["nodes"].as_array(),
-        pack["edges"].as_array(),
-        pack["routes"].as_array(),
-    ) {
-        let (coverage, _) = coverage_index(nodes, edges, routes);
-        if coverage != pack["coverage"] {
-            diagnostics.push(PackDiagnostic::error(
-                "coverage-index",
-                "/coverage",
-                "coverage index does not match the authority graph",
-            ));
-        }
-    }
-    diagnostics
-}
-
-pub fn inspect_pack_value(pack: &Value) -> Result<String, String> {
-    let statuses: BTreeMap<String, usize> = pack["nodes"]
-        .as_array()
-        .unwrap_or(&Vec::new())
-        .iter()
-        .fold(BTreeMap::new(), |mut counts, node| {
-            let status = node
-                .pointer("/verification/status")
-                .and_then(Value::as_str)
-                .unwrap_or("unknown")
-                .to_owned();
-            *counts.entry(status).or_default() += 1;
-            counts
-        });
-    Ok(stable_json(&json!({
-        "schema": "xinfa.context-pack-inspection/v1",
-        "project": pack["project"],
-        "cut": pack["cut"],
-        "visibility": pack["visibility"],
-        "roots": pack["roots"],
-        "counts": {
-            "sources": pack["inventory"].as_array().map(Vec::len).unwrap_or(0),
-            "nodes": pack["nodes"].as_array().map(Vec::len).unwrap_or(0),
-            "edges": pack["edges"].as_array().map(Vec::len).unwrap_or(0),
-            "routes": pack["routes"].as_array().map(Vec::len).unwrap_or(0),
-        },
-        "statuses": statuses,
-        "coverage": {
-            "claims": pack.pointer("/coverage/claims").and_then(Value::as_array).map(Vec::len).unwrap_or(0),
-            "orphans": pack.pointer("/coverage/orphans").cloned().unwrap_or_else(|| json!([])),
-        },
-        "diagnostics": pack["diagnostics"],
-    })))
-}
-
-pub fn verify_pack_artifacts(
-    pack_bytes: &[u8],
-    manifest_bytes: Option<&[u8]>,
-    receipt_bytes: Option<&[u8]>,
-) -> Result<(String, bool), String> {
-    let pack: Value = serde_json::from_slice(pack_bytes)
-        .map_err(|error| format!("invalid pack JSON: {error}"))?;
-    let mut diagnostics = verify_value(&pack);
-    match (manifest_bytes, receipt_bytes) {
-        (Some(manifest_bytes), Some(receipt_bytes)) => {
-            let manifest: Value = serde_json::from_slice(manifest_bytes)
-                .map_err(|error| format!("invalid manifest JSON: {error}"))?;
-            if manifest["schema"] != "xinfa.context-pack-manifest/v1"
-                || manifest["packRoot"] != pack["roots"]["pack"]
-                || manifest["artifacts"][0]["contentRoot"] != byte_digest(pack_bytes)
-                || manifest["artifacts"][0]["size"].as_u64() != Some(pack_bytes.len() as u64)
-            {
-                diagnostics.push(PackDiagnostic::error(
-                    "artifact-root",
-                    "/artifacts/0",
-                    "manifest does not bind the exact pack.json artifact",
-                ));
-            }
-            let mut manifest_core = manifest.clone();
-            manifest_core
-                .as_object_mut()
-                .and_then(|object| object.remove("manifestRoot"));
-            if digest(&manifest_core) != manifest["manifestRoot"].as_str().unwrap_or_default() {
-                diagnostics.push(PackDiagnostic::error(
-                    "manifest-root",
-                    "/manifestRoot",
-                    "manifest root does not match content",
-                ));
-            }
-            let receipt: Value = serde_json::from_slice(receipt_bytes)
-                .map_err(|error| format!("invalid receipt JSON: {error}"))?;
-            if receipt["schema"] != "xinfa.context-pack-compile-receipt/v1"
-                || receipt["verdict"] != "pass"
-                || receipt["packRoot"] != pack["roots"]["pack"]
-                || receipt["manifestRoot"] != manifest["manifestRoot"]
-                || receipt["qualifying"] != false
-                || receipt["selfCertified"] != false
-            {
-                diagnostics.push(PackDiagnostic::error(
-                    "receipt-binding",
-                    "/receipt",
-                    "compile receipt does not bind the verified Pack and manifest",
-                ));
-            }
-            let mut receipt_core = receipt.clone();
-            receipt_core
-                .as_object_mut()
-                .and_then(|object| object.remove("receiptRoot"));
-            if digest(&receipt_core) != receipt["receiptRoot"].as_str().unwrap_or_default() {
-                diagnostics.push(PackDiagnostic::error(
-                    "receipt-root",
-                    "/receiptRoot",
-                    "receipt root does not match content",
-                ));
-            }
-        }
-        (None, None) => {}
-        _ => {
-            return Err("pack verification requires both manifest and receipt artifacts".to_owned())
-        }
-    }
-    diagnostics.sort_by(|left, right| (&left.path, &left.code).cmp(&(&right.path, &right.code)));
-    let valid = diagnostics.is_empty();
-    let receipt = json!({
-        "schema": "xinfa.context-pack-verification-receipt/v1",
-        "valid": valid,
-        "qualifying": false,
-        "selfCertified": false,
-        "packRoot": pack.pointer("/roots/pack"),
-        "diagnostics": diagnostics_value(&diagnostics),
-    });
-    Ok((stable_json(&receipt), valid))
-}
-
-fn node_map(pack: &Value) -> BTreeMap<String, Value> {
-    pack["nodes"]
-        .as_array()
-        .unwrap_or(&Vec::new())
-        .iter()
-        .filter_map(|node| node["id"].as_str().map(|id| (id.to_owned(), node.clone())))
-        .collect()
-}
-
-fn changed_sources(old: &Value, new: &Value) -> BTreeSet<String> {
-    fn roots(value: &Value) -> BTreeMap<String, String> {
-        value["inventory"]
-            .as_array()
-            .unwrap_or(&Vec::new())
-            .iter()
-            .filter_map(|item| {
-                Some((
-                    item["path"].as_str()?.to_owned(),
-                    item["contentRoot"].as_str()?.to_owned(),
-                ))
-            })
-            .collect()
-    }
-    let old = roots(old);
-    let new = roots(new);
-    old.keys()
-        .chain(new.keys())
-        .filter(|path| old.get(*path) != new.get(*path))
-        .cloned()
-        .collect()
-}
-
-fn reverse_dependents(pack: &Value) -> BTreeMap<String, BTreeSet<String>> {
-    let mut reverse: BTreeMap<String, BTreeSet<String>> = BTreeMap::new();
-    for node in pack["nodes"].as_array().unwrap_or(&Vec::new()) {
-        let id = node["id"].as_str().unwrap_or_default();
-        for dependency in node
-            .pointer("/verification/dependencies")
-            .and_then(Value::as_array)
-            .unwrap_or(&Vec::new())
-        {
-            if let Some(target) = dependency["node"].as_str() {
-                reverse
-                    .entry(target.to_owned())
-                    .or_default()
-                    .insert(id.to_owned());
-            }
-        }
-    }
-    for edge in pack["edges"].as_array().unwrap_or(&Vec::new()) {
-        let from = edge["from"].as_str().unwrap_or_default();
-        let to = edge["to"].as_str().unwrap_or_default();
-        match edge["relation"].as_str().unwrap_or_default() {
-            "proves" => {
-                reverse
-                    .entry(from.to_owned())
-                    .or_default()
-                    .insert(to.to_owned());
-            }
-            "explains" | "defines" | "implements" | "depends-on" | "expands-to" => {
-                reverse
-                    .entry(to.to_owned())
-                    .or_default()
-                    .insert(from.to_owned());
-            }
-            _ => {}
-        }
-    }
-    reverse
-}
-
-pub fn impact_between_values(old: &Value, new_pack: &Value) -> Result<String, String> {
-    let changed_paths = changed_sources(old, new_pack);
-    let old_nodes = node_map(old);
-    let new_nodes = node_map(new_pack);
-    let mut seeds: BTreeSet<String> = old_nodes
-        .keys()
-        .chain(new_nodes.keys())
-        .filter(|id| old_nodes.get(*id) != new_nodes.get(*id))
-        .cloned()
-        .collect();
-    for (id, node) in old_nodes.iter().chain(new_nodes.iter()) {
-        if node
-            .pointer("/source/path")
-            .and_then(Value::as_str)
-            .is_some_and(|path| changed_paths.contains(path))
-        {
-            seeds.insert(id.clone());
-        }
-    }
-    let reverse = reverse_dependents(new_pack);
-    let mut affected = seeds.clone();
-    let mut queue: VecDeque<String> = seeds.into_iter().collect();
-    while let Some(id) = queue.pop_front() {
-        for dependent in reverse.get(&id).into_iter().flatten() {
-            if affected.insert(dependent.clone()) {
-                queue.push_back(dependent.clone());
-            }
-        }
-    }
-    let kinds: BTreeMap<String, String> = old_nodes
-        .iter()
-        .chain(new_nodes.iter())
-        .filter_map(|(id, node)| Some((id.clone(), node["kind"].as_str()?.to_owned())))
-        .collect();
-    let select_kind = |kind: &str| -> Vec<String> {
-        affected
-            .iter()
-            .filter(|id| {
-                kinds
-                    .get(*id)
-                    .is_some_and(|value| value == kind || (kind == "claim" && value == "invariant"))
-            })
-            .cloned()
-            .collect()
-    };
-    let affected_routes: Vec<String> = new_pack["routes"]
-        .as_array()
-        .unwrap_or(&Vec::new())
-        .iter()
-        .filter(|route| {
-            route["nodes"]
-                .as_array()
-                .unwrap_or(&Vec::new())
-                .iter()
-                .any(|id| id.as_str().is_some_and(|id| affected.contains(id)))
-        })
-        .filter_map(|route| route["id"].as_str().map(str::to_owned))
-        .collect();
-    Ok(stable_json(&json!({
-        "schema": "xinfa.context-pack-impact/v1",
-        "since": old.pointer("/roots/pack"),
-        "current": new_pack.pointer("/roots/pack"),
-        "changedSources": changed_paths,
-        "affectedNodes": affected,
-        "affectedClaims": select_kind("claim"),
-        "affectedDocuments": select_kind("document"),
-        "affectedRoutes": affected_routes,
-    })))
 }
 
 pub fn pack_value(artifacts: &PackArtifacts) -> Result<Value, String> {
